callback refactor
