+++ libs/langchain/langchain/callbacks/tracers/__init__.py
@@ -1,12 +1,12 @@
 """Tracers that record execution of LangChain runs."""
 
-from langchain.callbacks.tracers.langchain import LangChainTracer
-from langchain.callbacks.tracers.langchain_v1 import LangChainTracerV1
-from langchain.callbacks.tracers.stdout import (
+from langchain.callbacks.tracers.wandb import WandbTracer
+from langchain.schema.callbacks.tracers.langchain import LangChainTracer
+from langchain.schema.callbacks.tracers.langchain_v1 import LangChainTracerV1
+from langchain.schema.callbacks.tracers.stdout import (
     ConsoleCallbackHandler,
     FunctionCallbackHandler,
 )
-from langchain.callbacks.tracers.wandb import WandbTracer
 
 __all__ = [
     "LangChainTracer",
